Minor changes to home screen UI
