+++ app/src/main/java/com/simpletasks/ui/components/TaskListComponent.kt
@@ -1,7 +1,6 @@
 package com.simpletasks.ui.components
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Checkbox
@@ -16,7 +15,7 @@ import androidx.compose.ui.unit.dp
 @Composable
 fun TaskListComponent(listOfTasks: List<String>, modifier: Modifier = Modifier) {
     if (listOfTasks.isEmpty()) {
-        Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+        Box(modifier = modifier, contentAlignment = Alignment.Center) {
             Text(text = "No Tasks!")
         }
     } else {
