+++ app/src/main/java/com/simpletasks/ui/components/TopAppBarComponent.kt
@@ -1,0 +1,30 @@
+package com.simpletasks.ui.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun TopAppBar(modifier: Modifier = Modifier) {
+    Row(
+        modifier = modifier
+            .padding(top = 16.dp, bottom = 16.dp, start = 8.dp, end = 8.dp)
+            .fillMaxWidth()
+            .background(MaterialTheme.colorScheme.background),
+        horizontalArrangement = Arrangement.Center,
+    ) {
+        Text(
+            text = "Simple Tasks",
+            textAlign = TextAlign.Center,
+            style = MaterialTheme.typography.titleLarge
+        )
+    }
+}
